🚚 split ag-ui-stream and its types into per-domain files and folders
ag-ui-stream.ts (356 lines) and ag-ui-stream.types.ts (190 lines) mixed run,
message, tool, A2UI surface, and Agent-thread-delivery lifecycle handling in
two files. Each domain now owns its own folder with a paired *.types.ts (and
a *.validator.ts next to it where runtime validation applies), following the
same layout already used by libs/backend/agents/personal/{personas,
configuration}. The public barrel and package behavior are unchanged; all 19
existing tests pass unmodified against the split implementation.
