+++ libs/frontend/state/conversation/ag-ui/src/__tests__/ag-ui-stream.test.ts
@@ -4,8 +4,11 @@ import { describe, expect, it } from "vitest";
 import { AG_UI_A2UI_ENVELOPE_VERSION, AG_UI_AGENT_THREAD_PARENT_DELIVERY_EVENT, AG_UI_INTERRUPTS_CLEARED_EVENT, AG_UI_TOOL_FAILURE_EVENT, AG_UI_TOOL_RECOVERY_REQUIRED_EVENT, AgentThreadDeliveryKinds, AgUiA2uiSurfaceStates, AgUiToolRecoveryProviderOutcomes } from "@opencrane/contracts";
 
 import { __DecodeAgUiSseRecord } from "../ag-ui-sse-decoder.js";
-import { AgUiMessageStatuses, AgUiRunStatuses, AgUiToolStatuses, type AgUiStreamRecord } from "../ag-ui-stream.types.js";
 import { __AgUiResumeCursor, __CreateAgUiStreamState, __ReduceAgUiStream } from "../ag-ui-stream.js";
+import type { AgUiStreamRecord } from "../ag-ui-stream.types.js";
+import { AgUiMessageStatuses } from "../message/message.types.js";
+import { AgUiRunStatuses } from "../run/run.types.js";
+import { AgUiToolStatuses } from "../tool/tool.types.js";
 
 /** Exact safe progressive-disclosure fixture. */
 function _ToolFailureEnvelope(failureCode = "AuthenticationError", retrying = true)
