+++ libs/frontend/state/conversation/ag-ui/src/agent-thread-delivery/agent-thread-delivery.types.ts
@@ -1,0 +1,4 @@
+import type { AgUiAgentThreadParentDeliveryEnvelope } from "@opencrane/contracts";
+
+/** One validated display-safe Agent-thread delivery projected into an immediate parent stream. */
+export type AgUiAgentThreadParentDelivery = AgUiAgentThreadParentDeliveryEnvelope;
