+++ libs/frontend/state/conversation/ag-ui/src/index.ts
@@ -1,3 +1,7 @@
 export { __DecodeAgUiSseRecord } from "./ag-ui-sse-decoder.js";
 export { __AgUiResumeCursor, __CreateAgUiStreamState, __ReduceAgUiStream, __RevokeAgUiStreamAccess } from "./ag-ui-stream.js";
-export { AgUiMessageStatuses, AgUiRunStatuses, AgUiToolStatuses, type AgUiAgentThreadParentDelivery, type AgUiMessageView, type AgUiRunFailure, type AgUiStreamRecord, type AgUiStreamState, type AgUiToolFailure, type AgUiToolView } from "./ag-ui-stream.types.js";
+export type { AgUiStreamRecord, AgUiStreamState } from "./ag-ui-stream.types.js";
+export type { AgUiAgentThreadParentDelivery } from "./agent-thread-delivery/agent-thread-delivery.types.js";
+export { AgUiMessageStatuses, type AgUiMessageView } from "./message/message.types.js";
+export { AgUiRunStatuses, type AgUiRunFailure } from "./run/run.types.js";
+export { AgUiToolStatuses, type AgUiToolFailure, type AgUiToolView } from "./tool/tool.types.js";
